Backend change to support custom form (#36)
* Event Form schema

* Make GET API + add event form to event schema

* Change question schema, remove ID

* Handle error

* add new field types to EventForm, add eventFormAnswers object to Attendee

* addEventForm endpoint, with validation

* pull from staging

* implement custom form except file upload

* implement file upload for registration

---------
